+++ cmd/aumo/main.go
@@ -9,9 +9,9 @@ import (
 	"github.com/deliriumproducts/aumo/auth"
 	"github.com/deliriumproducts/aumo/mysql"
 	"github.com/deliriumproducts/aumo/net/http/rest"
-	"github.com/deliriumproducts/aumo/product"
+	"github.com/deliriumproducts/aumo/products"
 	"github.com/deliriumproducts/aumo/receipt"
-	"github.com/deliriumproducts/aumo/user"
+	"github.com/deliriumproducts/aumo/users"
 	"github.com/gomodule/redigo/redis"
 	"github.com/joho/godotenv"
 	upper "upper.io/db.v3/mysql"
