+++ users/service.go
@@ -1,4 +1,4 @@
-package user
+package users
 
 import "github.com/deliriumproducts/aumo"
